delete unnecessary files, rename gasmaleperf to jho; use separate file to generate perfance output
