--- model/sketch_params.csv
@@ -1,16 +1,0 @@
-,Value,Units,Label
-L,5.49999999981,ft,tail boom length
-S,24.8841087548,ft**2,wing area
-SM,0.18536240741976534,-,static margin
-S_h,1.36793256131,ft**2,horizontal tail area
-S_v,2.21415713402,ft**2,total vertical tail surface area
-b,24.4691589252,ft,Span
-b_h,2.61527490076,ft,horizontal tail span
-b_v,1.48800441331,ft,vertical tail span
-d,0.984745021863,ft,fuselage diameter
-d_0,0.08209969059,ft,tail boom diameter
-eta,3,-,tail boom separation/fuselage diameter
-l_{fuel},2.83995516858,ft,fuel tank length
-lambda,0.5,-,taper ratio
-x_{cg},1.9225562228132831,foot,center of gravity
-x_{np},2.1110620198205488,foot,neutral point
